* Fixed: Bug in Previous commit.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@6420 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/wxscintilla/include/wx/wxscintilla.h
@@ -2227,7 +2227,7 @@ class  WXDLLIMPEXP_SCI wxScintillaEvent;
 
 /* C::B begin */
 #if !wxCHECK_VERSION(2, 9, 0)
-typedef size_t wxIntPtr;
+typedef long wxIntPtr;
 #endif
 
 typedef wxIntPtr (* wxSciFnDirect) (wxIntPtr ptr, unsigned int iMessage,
